Update the IAM Policy chart to account for the operator-sdk upgrade (#53)
After the operator-sdk was upgraded for the IAM Policy Controller, the
privileges required were reduced and a leader election role was added.
This leader election role is necessary because the IAM Policy Controller
either requires access to ConfigMaps for the legacy leader election in
OCP 3.11 or the lease API for OCP 4.x+.

An additional argument is also passed to the controller when it is running
on OCP 3.11 in order to use the legacy leader election method of
ConfigMaps.

Relates to:
https://github.com/open-cluster-management/iam-policy-controller/pull/78

Signed-off-by: mprahl <mprahl@users.noreply.github.com>