GRC RBAC updates (#79)
* Align template file names

The names within the GRC components now match the names of the other
components. For example, they use:
- <group>_<kind>_crd.yaml
- deployment.yaml
- service_account.yaml
- role_binding.yaml
- cluster_role.yaml
- leader_election_role.yaml

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>

* Update cert-policy-controller rbac

The namespaced role had many more permissions than it needed. Based on
the resources in the controller repo, it only needed the leader election
rbac.

The updated cluster role matches the updates in
https://github.com/stolostron/cert-policy-controller/pull/122

Refs:
 - https://github.com/open-cluster-management/backlog/issues/17918

Signed-off-by: Justin Kulikauskas <jkulikau@redhat.com>